Test.Golden.Sophie.Key.ConvertBccAddressKey

Documentation

golden_convertBccAddressColeSigningKey ∷ Property

Test that converting a bcc-address Cole signing key yields the expected result.

golden_convertBccAddressIcarusSigningKey ∷ Property

Test that converting a bcc-address Icarus signing key yields the expected result.

golden_convertBccAddressSophiePaymentSigningKey ∷ Property

Test that converting a bcc-address Sophie payment signing key yields the expected result.

golden_convertBccAddressSophieStakeSigningKey ∷ Property

Test that converting a bcc-address Sophie stake signing key yields the expected result.